+++ test/index.js
@@ -8,7 +8,7 @@ require('./cli/jsdoc-md.test')(tests);
 require('./private/deconstructJsdocNamepath.test')(tests);
 require('./private/getJsdocAstTag.test')(tests);
 require('./private/getJsdocAstTags.test')(tests);
-require('./private/jsdocCommentsFromCode.test')(tests);
+require('./private/codeToJsdocComments.test')(tests);
 require('./private/jsdocCommentToMember.test')(tests);
 require('./private/mdFileReplaceSection.test')(tests);
 require('./private/mdToMdAst.test')(tests);